FastAPIMongoEngineGraphQL/app/dev_requirements.txt
fastapi
uvicorn
gunicorn
pandas==1.0.3
numpy==1.18.2
psycopg2==2.8.5
sqlalchemy==1.3.17
jwt==1.0.0
bcrypt==3.1.7
passlib==1.7.2
python-jose
python-multipart
async-exit-stack
async-generator
async-exit-stack
async-generator
uuid==1.30
graphene-mongo